Add buttons for undo/redo functionality
